Styling show

+++ app/javascript/react/components/ProjectShow.js
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from "react"
 import { Redirect } from "react-router-dom"
 import _, { set } from "lodash"
 import CommentTile from "./CommentTile"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+import { faTimesCircle } from "@fortawesome/free-solid-svg-icons"
 
 const ProjectShow = props => {
   const { id } = props.match.params
@@ -30,10 +32,13 @@ const ProjectShow = props => {
   }
 
   const deleteFunc = async () => {
-    await fetch(`/api/v1/projects/${project.id}`, {
-      method: "DELETE",
-    })
-    setDeleteRedirect(true)
+    let bool = confirm("Are you sure you want to delete this project?")
+    if(bool === true) {
+      await fetch(`/api/v1/projects/${project.id}`, {
+        method: "DELETE",
+      })
+      setDeleteRedirect(true)
+    }
   }
 
   const postJoin = async event => {
@@ -61,13 +66,17 @@ const ProjectShow = props => {
       credentials: "same-origin",
       body: JSON.stringify({project: {id: id}, user: {id: userId}})
     })
-    const responseBody = response.json()
-    if(responseBody.response == "User joined successfully") {
-      setUserRole("member")
+    const responseBody = await response.json()
+    if(responseBody.response == "User added successfully") {
+      setProject({
+        ...project,
+        users: project.users.concat(responseBody.user),
+        requests: project.requests.filter(user => user.id !== responseBody.user.id)
+      })
     }
   }
 
-  if(deleteRedirect) {
+  if(deleteRedirect) {  
     return <Redirect to="/my_projects" />
   }
 
@@ -75,9 +84,20 @@ const ProjectShow = props => {
     return <Redirect to={`/projects/${project.id}/edit`} />
   }
 
-  let button  
+  let joinButton  
   if(userRole == "viewer") {
-    button = <button onClick={postJoin} className="button">JOIN</button>
+    joinButton = <button onClick={postJoin} className="show-button">JOIN</button>
+  }
+
+  let editButton
+  let deleteButton
+  if(userRole == "owner") {
+    editButton = <button className="show-button" onClick={editFunc}>
+                  Edit
+                 </button> 
+    deleteButton = <button className="show-button" onClick={deleteFunc}>
+                    Delete
+                   </button>
   }
   
   let requests
@@ -88,7 +108,7 @@ const ProjectShow = props => {
       }
       return <button 
               onClick={acceptRequest} 
-              className="button">
+              className="request-tile">
                 {request.name}
               </button>
     })
@@ -150,48 +170,53 @@ const ProjectShow = props => {
     setComment(event.currentTarget.value)
   }
 
+
+  const closeFlash = () => {
+    setJoinResponse(null)
+  }
+
+  let flashMessage
+  if(joinResponse) {
+    flashMessage = <p className="flash-container">{joinResponse} <FontAwesomeIcon onClick={closeFlash} className="x-button" size="lg" icon={faTimesCircle} /></p>
+  }
+
   return (
-    <div className="grid-container gray">
-      <div className="black-background">
-        {error}
-        {joinResponse}
-        <div className="show-header-container">
-          <h1 className="project-show-title">
-            {project.name} 
-          </h1>
-          <button className="show-button" onClick={deleteFunc}>
-              Delete
-          </button>
-          <button className="show-button" onClick={editFunc}>
-            Edit
-          </button>
-          {button}
-          <p className="project-show-desc">{project.description}</p>
+    <div>
+      {deleteButton}
+      {editButton}
+      {joinButton}
+      <div className="show-margin">
+      {error}
+      {flashMessage}
+      <div className="show-header-container">
+        <h1 className="project-show-title">
+          {project.name} 
+        </h1>
+        <p className="project-show-desc">{project.description}</p>
+      </div>
+      <div className="text-left grid-x grid-margin-x">
+        <div className="cell small-9">
+          <h4 className="show-sub-header">COMMENTS</h4>
+          <form onSubmit={handleSubmit}>
+            <textarea 
+              className="project-text-area wide" 
+              name="description" 
+              placeholder="Leave a comment..." 
+              value={comment} 
+              onChange={handleChange}>
+            </textarea>
+            <input type="submit" value="Comment" className="show-submit-button" />
+          </form>
+          {comments}
         </div>
-        <div className="text-left grid-x grid-margin-x">
-          <div className="cell small-9">
-            <h4 className="show-sub-header">COMMENTS</h4>
-            <form onSubmit={handleSubmit}>
-              <textarea 
-                className="project-text-area wide" 
-                name="description" 
-                placeholder="Leave a comment..." 
-                value={comment} 
-                onChange={handleChange}>
-              </textarea>
-              <input type="submit" value="Comment" className="show-submit-button" />
-            </form>
-            {comments}
-          </div>
-          <div className="text-center cell small-3">
-            <h4 className="show-sub-header">MEMBERS</h4>
-            {userTiles}
-          </div>
+        <div className="text-center cell small-3">
+          <h4 className="show-sub-header">MEMBERS</h4>
           {requests}
+          {userTiles}
         </div>
       </div>
     </div>
-
+    </div>
   )
 }
